fix(math-evidence): cast operator ID to u8 for registry.dispatch — fix broken 5c6a9d1b

+++ backend/searcher-rs/src/math_evidence.rs
@@ -100,7 +100,7 @@ pub fn evaluate_strategy_operators(
     operator_ids
         .iter()
         .filter_map(|&id| {
-            let out = registry.dispatch(id, state)?;
+            let out = registry.dispatch(id as u8, state)?;
             Some((id, out.scalar_value, out.operator_name))
         })
         .collect()
